fix(server): tool refactor
