+++ protocol/go.mod
@@ -2,4 +2,7 @@ module github.com/latebit/demarkus/protocol
 
 go 1.26
 
-require gopkg.in/yaml.v3 v3.0.1
+require (
+	github.com/BurntSushi/toml v1.6.0
+	gopkg.in/yaml.v3 v3.0.1
+)
