+++ server/cmd/demarkus-token/main.go
@@ -1,8 +1,16 @@
+// demarkus-token mints, lists, and revokes capability tokens for a Mark
+// Protocol server. The TOML file it writes is consumed by the server's
+// auth.LoadTokens at startup (and on SIGHUP).
+//
+// Usage:
+//
+//	demarkus-token generate -label fritz-laptop -paths "/docs/*" -ops "publish" -tokens tokens.toml
+//	demarkus-token list -tokens tokens.toml
+//	demarkus-token revoke -label fritz-laptop -tokens tokens.toml
 package main
 
 import (
-	"crypto/rand"
-	"encoding/hex"
+	"errors"
 	"flag"
 	"fmt"
 	"log"
@@ -10,8 +18,7 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/BurntSushi/toml"
-	"github.com/latebit/demarkus/protocol"
+	"github.com/latebit/demarkus/protocol/token"
 )
 
 func main() {
@@ -62,69 +69,27 @@ func cmdGenerate(args []string) {
 		os.Exit(1)
 	}
 
-	// Generate 32 random bytes → 64 hex chars.
-	secret := make([]byte, 32)
-	if _, err := rand.Read(secret); err != nil {
-		log.Fatalf("generate random bytes: %v", err)
+	minted, err := token.Generate(*label, splitTrimmed(*paths), splitTrimmed(*ops))
+	if err != nil {
+		log.Fatalf("generate: %v", err)
 	}
-	rawToken := hex.EncodeToString(secret)
-	hashedToken := protocol.HashToken(rawToken)
-
-	pathList := splitTrimmed(*paths)
-	opsList := splitTrimmed(*ops)
-
-	// Format TOML entry.
-	entry := fmt.Sprintf("\n[tokens.%s]\nhash = %q\npaths = [%s]\noperations = [%s]\n",
-		*label,
-		hashedToken,
-		quotedList(pathList),
-		quotedList(opsList),
-	)
 
 	if *tokensFile != "" {
-		// Check for duplicate label in existing file.
-		var tf tokensFileData
-		if _, err := os.Stat(*tokensFile); err == nil {
-			if _, err := toml.DecodeFile(*tokensFile, &tf); err != nil {
-				log.Fatalf("read tokens file: %v", err)
+		if err := token.AppendEntry(*tokensFile, minted.Label, &minted.Entry); err != nil {
+			if errors.Is(err, token.ErrLabelExists) {
+				log.Fatalf("label %q already exists in %s", minted.Label, *tokensFile)
 			}
-			if _, exists := tf.Tokens[*label]; exists {
-				log.Fatalf("label %q already exists in %s", *label, *tokensFile)
-			}
+			log.Fatalf("%v", err)
 		}
-
-		f, err := os.OpenFile(*tokensFile, os.O_APPEND|os.O_CREATE|os.O_WRONLY, 0o600)
-		if err != nil {
-			log.Fatalf("open tokens file: %v", err)
-		}
-		if _, err := f.WriteString(entry); err != nil {
-			_ = f.Close()
-			log.Fatalf("write token entry: %v", err)
-		}
-		if err := f.Close(); err != nil {
-			log.Fatalf("close tokens file: %v", err)
-		}
-		fmt.Fprintf(os.Stderr, "Token %q appended to %s\n", *label, *tokensFile)
+		fmt.Fprintf(os.Stderr, "Token %q appended to %s\n", minted.Label, *tokensFile)
 	} else {
 		fmt.Fprintln(os.Stderr, "Add this to your tokens.toml:")
-		fmt.Fprint(os.Stderr, entry)
+		fmt.Fprint(os.Stderr, token.FormatEntry(minted.Label, &minted.Entry))
 	}
 
 	fmt.Fprintln(os.Stderr)
 	fmt.Fprintln(os.Stderr, "Raw token (give to client, shown once):")
-	fmt.Println(rawToken)
-}
-
-// tokensFileData is the TOML structure for reading the full file.
-type tokensFileData struct {
-	Tokens map[string]tokenEntry `toml:"tokens"`
-}
-
-type tokenEntry struct {
-	Hash       string   `toml:"hash"`
-	Paths      []string `toml:"paths"`
-	Operations []string `toml:"operations"`
-	Expires    string   `toml:"expires,omitempty"`
+	fmt.Println(minted.Raw)
 }
 
 func cmdList(args []string) {
@@ -143,23 +108,23 @@ func cmdList(args []string) {
 		os.Exit(1)
 	}
 
-	var tf tokensFileData
-	if _, err := toml.DecodeFile(*tokensFile, &tf); err != nil {
-		log.Fatalf("read tokens file: %v", err)
+	file, err := token.ReadFile(*tokensFile)
+	if err != nil {
+		log.Fatalf("%v", err)
 	}
 
-	if len(tf.Tokens) == 0 {
+	if len(file.Tokens) == 0 {
 		fmt.Println("No tokens found.")
 		return
 	}
 
-	labels := make([]string, 0, len(tf.Tokens))
-	for label := range tf.Tokens {
+	labels := make([]string, 0, len(file.Tokens))
+	for label := range file.Tokens {
 		labels = append(labels, label)
 	}
 	sort.Strings(labels)
 	for _, label := range labels {
-		tok := tf.Tokens[label]
+		tok := file.Tokens[label]
 		paths := strings.Join(tok.Paths, ", ")
 		ops := strings.Join(tok.Operations, ", ")
 		fmt.Printf("%-20s  paths: %-20s  ops: %s\n", label, paths, ops)
@@ -183,35 +148,17 @@ func cmdRevoke(args []string) {
 		os.Exit(1)
 	}
 
-	var tf tokensFileData
-	if _, err := toml.DecodeFile(*tokensFile, &tf); err != nil {
-		log.Fatalf("read tokens file: %v", err)
+	file, err := token.ReadFile(*tokensFile)
+	if err != nil {
+		log.Fatalf("%v", err)
 	}
-
-	if _, ok := tf.Tokens[*label]; !ok {
+	if _, ok := file.Tokens[*label]; !ok {
 		log.Fatalf("token %q not found in %s", *label, *tokensFile)
 	}
+	delete(file.Tokens, *label)
 
-	delete(tf.Tokens, *label)
-
-	// Atomic write: write to temp file, then rename to avoid truncating on failure.
-	tmpFile := *tokensFile + ".tmp"
-	f, err := os.OpenFile(tmpFile, os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0o600)
-	if err != nil {
-		log.Fatalf("open temp tokens file: %v", err)
-	}
-	if err := toml.NewEncoder(f).Encode(tf); err != nil {
-		_ = f.Close()
-		_ = os.Remove(tmpFile)
-		log.Fatalf("write tokens file: %v", err)
-	}
-	if err := f.Close(); err != nil {
-		_ = os.Remove(tmpFile)
-		log.Fatalf("close tokens file: %v", err)
-	}
-	if err := os.Rename(tmpFile, *tokensFile); err != nil {
-		_ = os.Remove(tmpFile)
-		log.Fatalf("rename tokens file: %v", err)
+	if err := token.WriteFile(*tokensFile, file); err != nil {
+		log.Fatalf("%v", err)
 	}
 
 	fmt.Fprintf(os.Stderr, "Revoked token %q\n", *label)
@@ -227,12 +174,4 @@ func splitTrimmed(s string) []string {
 		}
 	}
 	return out
-}
-
-func quotedList(items []string) string {
-	quoted := make([]string, len(items))
-	for i, item := range items {
-		quoted[i] = fmt.Sprintf("%q", item)
-	}
-	return strings.Join(quoted, ", ")
 }
