+++ server/go.mod
@@ -3,17 +3,17 @@ module github.com/latebit/demarkus/server
 go 1.26
 
 require (
+	github.com/BurntSushi/toml v1.6.0
 	github.com/latebit/demarkus/protocol v0.0.0
 	github.com/quic-go/quic-go v0.59.0
+	golang.org/x/time v0.14.0
 )
 
 require (
-	github.com/BurntSushi/toml v1.6.0 // indirect
 	github.com/kr/text v0.2.0 // indirect
 	golang.org/x/crypto v0.41.0 // indirect
 	golang.org/x/net v0.43.0 // indirect
 	golang.org/x/sys v0.35.0 // indirect
-	golang.org/x/time v0.14.0 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
